brought in named Sphere Calculation data

+++ src/Shared/FogentRoleplayLib.fs
@@ -502,8 +502,8 @@ module Range =
 module AreaOfEffectCalculation =
 
     type SphereCalculation = {
-        initRadius: uint
-        radiusPerDice: uint
+        initRadius: float
+        radiusPerDice: float
     }
 
     type ConeCalculation = {
